requirements.txt
# Core
pandas>=2.0
numpy>=1.24
pydantic>=2.0
pydantic-settings>=2.0

# Data fetching
yfinance>=0.2.30
requests>=2.31
feedparser>=6.0
aiohttp>=3.9
firecrawl-py>=1.0

# ML
scikit-learn>=1.3
xgboost>=2.0
lightgbm>=4.0

# Technical Analysis
ta>=0.11

# LLM
openai>=1.0

# Storage
duckdb>=0.9

# UI
streamlit>=1.28

# Utils
python-dotenv>=1.0
loguru>=0.7
tenacity>=8.2
pytz>=2023.3
